Feat: delete asset api
