+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+15",
+  "version": "0.8.5+16",
   "private": false,
   "scripts": {
     "dev": "next dev",
